pcle: clean up specie.cxx

+++ libs/simul/pcle/specie.cxx
@@ -22,7 +22,7 @@ int Specie<T,D>::Clean (int *clean)
 {
   int cleaned = 0;
 
-  size_t lcmd = _cmdqueue.GetSize ();
+  int lcmd = _cmdqueue.GetSize ();
   PcleCommandInfo *linfo;
 
   if (lcmd == 0)
@@ -48,14 +48,14 @@ int Specie<T,D>::Clean (int *clean)
 
       if (linfo->pid == rmvd)
       {
-	info.cmd = linfo->cmd;
-	linfo->cmd = PCLE_CMD_NONE;
+      	info.cmd = linfo->cmd;
+      	linfo->cmd = PCLE_CMD_NONE;
 
-	do
-	{
-	  linfo = &(_cmdqueue.Get (--lcmd));
-	}
-	while ((!(linfo->cmd) || (linfo->cmd & PCLE_CMD_REMOVE)) && lcmd>0);
+      	do
+      	{
+      	  linfo = &(_cmdqueue.Get (--lcmd));
+      	}
+      	while ((!(linfo->cmd) || (linfo->cmd & PCLE_CMD_REMOVE)) && lcmd>0);
       }
 
       ++cleaned;
